resume-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain-ollama
langchain-community
pymupdf
pydantic
python-dotenv